Method matching file name
