+++ Environment_Engine/Modify/SplitOpeningByGeometry.cs
